+++ interceptors/interceptors-now/src/main/java/org/supertribe/interceptors/interim/ImageBean.java
@@ -16,24 +16,10 @@
  */
 package org.supertribe.interceptors.interim;
 
-import javax.interceptor.AroundInvoke;
-import javax.interceptor.Interceptor;
-import javax.interceptor.InvocationContext;
-
-import static org.supertribe.interceptors.interim.Utils.subtractTwo;
-import static org.supertribe.interceptors.interim.Utils.wrapResult;
-
 /**
  * @version $Rev$ $Date$
  */
-@SuperclassLevelInterceptor
-@Interceptor
-public class ClassLevelInterceptorSuperClassOne {
-
-    @AroundInvoke
-    public Object businessMethodInterceptor(final InvocationContext ic) throws Exception {
-        subtractTwo(ic);
-        return wrapResult(ic, this.getClass().getSimpleName());
-    }
+@Orange
+public class ImageBean {
 
 }
