+++ interceptors/interceptors-now/src/main/java/org/supertribe/interceptors/interim/Orange.java
@@ -11,5 +11,5 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 @Inherited
 @Target({ElementType.TYPE})
-public @interface ClassLevelInterceptor {
+public @interface Orange {
 }
