+++ interceptors/interceptors-now/src/main/java/org/supertribe/interceptors/interim/Red.java
@@ -10,6 +10,6 @@ import java.lang.annotation.Target;
 @InterceptorBinding
 @Retention(RetentionPolicy.RUNTIME)
 @Inherited
-@Target({ElementType.TYPE, ElementType.METHOD})
+@Target({ElementType.TYPE})
 public @interface Red {
 }
