+++ interceptors/interceptors-now/src/main/java/org/supertribe/interceptors/interim/RedInterceptor.java
@@ -20,7 +20,7 @@ import javax.interceptor.AroundInvoke;
 import javax.interceptor.Interceptor;
 import javax.interceptor.InvocationContext;
 
-import static org.supertribe.interceptors.interim.Utils.subtractTwo;
+import static org.supertribe.interceptors.interim.Utils.subtractThree;
 import static org.supertribe.interceptors.interim.Utils.wrapResult;
 
 /**
@@ -32,7 +32,7 @@ public class RedInterceptor {
 
     @AroundInvoke
     public Object businessMethodInterceptor(final InvocationContext ic) throws Exception {
-        subtractTwo(ic);
+        subtractThree(ic);
         return wrapResult(ic, "Red");
     }
