--- interceptors/interceptors-now/src/main/java/org/supertribe/interceptors/interim/SuperclassLevelInterceptor.java
@@ -1,15 +1,0 @@
-package org.supertribe.interceptors.interim;
-
-import javax.interceptor.InterceptorBinding;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Inherited;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@InterceptorBinding
-@Retention(RetentionPolicy.RUNTIME)
-@Inherited
-@Target({ElementType.TYPE})
-public @interface SuperclassLevelInterceptor {
-}
